+++ price-feeder/oracle/oracle.go
@@ -34,10 +34,6 @@ const (
 	tickerSleep = 1000 * time.Millisecond
 )
 
-// deviationThreshold defines how many 𝜎 a provider can be away from the mean
-// without being considered faulty.
-var deviationThreshold = sdk.MustNewDecFromStr("1.0")
-
 // PreviousPrevote defines a structure for defining the previous prevote
 // submitted on-chain.
 type PreviousPrevote struct {
@@ -413,281 +409,6 @@ func NewProvider(ctx context.Context, providerName string, logger zerolog.Logger
 	}
 
 	return nil, fmt.Errorf("provider %s not found", providerName)
-}
-
-// FilterTickerDeviations finds the standard deviations of the prices of
-// all assets, and filters out any providers that are not within 2𝜎 of the mean.
-func FilterTickerDeviations(
-	logger zerolog.Logger,
-	prices provider.AggregatedProviderPrices,
-) (provider.AggregatedProviderPrices, error) {
-	var (
-		filteredPrices = make(provider.AggregatedProviderPrices)
-		priceMap       = make(map[string]map[string]sdk.Dec)
-	)
-
-	for providerName, priceTickers := range prices {
-		if _, ok := priceMap[providerName]; !ok {
-			priceMap[providerName] = make(map[string]sdk.Dec)
-		}
-		for base, tp := range priceTickers {
-			priceMap[providerName][base] = tp.Price
-		}
-	}
-
-	deviations, means, err := StandardDeviation(priceMap)
-	if err != nil {
-		return nil, err
-	}
-
-	// accept any prices that are within 2𝜎, or for which we couldn't get 𝜎
-	for providerName, priceTickers := range prices {
-		for base, tp := range priceTickers {
-			if _, ok := deviations[base]; !ok ||
-				(tp.Price.GTE(means[base].Sub(deviations[base].Mul(deviationThreshold))) &&
-					tp.Price.LTE(means[base].Add(deviations[base].Mul(deviationThreshold)))) {
-				if _, ok := filteredPrices[providerName]; !ok {
-					filteredPrices[providerName] = make(map[string]provider.TickerPrice)
-				}
-
-				filteredPrices[providerName][base] = tp
-			} else {
-				telemetry.IncrCounter(1, "failure", "provider", "type", "ticker")
-				logger.Warn().
-					Str("base", base).
-					Str("provider", providerName).
-					Str("price", tp.Price.String()).
-					Msg("provider deviating from other prices")
-			}
-		}
-	}
-
-	return filteredPrices, nil
-}
-
-// FilterCandleDeviations finds the standard deviations of the tvwaps of
-// all assets, and filters out any providers that are not within 2𝜎 of the mean.
-func FilterCandleDeviations(
-	logger zerolog.Logger,
-	candles provider.AggregatedProviderCandles,
-) (provider.AggregatedProviderCandles, error) {
-	var (
-		filteredCandles = make(provider.AggregatedProviderCandles)
-		tvwaps          = make(map[string]map[string]sdk.Dec)
-	)
-
-	for providerName, priceCandles := range candles {
-		candlePrices := make(provider.AggregatedProviderCandles)
-
-		for base, cp := range priceCandles {
-			if _, ok := candlePrices[providerName]; !ok {
-				candlePrices[providerName] = make(map[string][]provider.CandlePrice)
-			}
-
-			candlePrices[providerName][base] = cp
-		}
-
-		tvwap, err := ComputeTVWAP(candlePrices)
-		if err != nil {
-			return nil, err
-		}
-
-		for base, asset := range tvwap {
-			if _, ok := tvwaps[providerName]; !ok {
-				tvwaps[providerName] = make(map[string]sdk.Dec)
-			}
-
-			tvwaps[providerName][base] = asset
-		}
-	}
-
-	deviations, means, err := StandardDeviation(tvwaps)
-	if err != nil {
-		return nil, err
-	}
-
-	// accept any tvwaps that are within 2𝜎, or for which we couldn't get 𝜎
-	for providerName, priceMap := range tvwaps {
-		for base, price := range priceMap {
-			if _, ok := deviations[base]; !ok ||
-				(price.GTE(means[base].Sub(deviations[base].Mul(deviationThreshold))) &&
-					price.LTE(means[base].Add(deviations[base].Mul(deviationThreshold)))) {
-				if _, ok := filteredCandles[providerName]; !ok {
-					filteredCandles[providerName] = make(map[string][]provider.CandlePrice)
-				}
-
-				filteredCandles[providerName][base] = candles[providerName][base]
-			} else {
-				telemetry.IncrCounter(1, "failure", "provider", "type", "candle")
-				logger.Warn().
-					Str("base", base).
-					Str("provider", providerName).
-					Str("price", price.String()).
-					Msg("provider deviating from other candles")
-			}
-		}
-	}
-
-	return filteredCandles, nil
-}
-
-// ConvertCandlesToUSD converts any candles which are not quoted in USD
-// to USD by other price feeds.
-func convertCandlesToUSD(
-	candles provider.AggregatedProviderCandles,
-	providerPairs map[string][]types.CurrencyPair,
-) (provider.AggregatedProviderCandles, error) {
-	if len(candles) == 0 {
-		return candles, nil
-	}
-
-	conversionRates := make(map[string]sdk.Dec)
-	requiredConversions := make(map[string]types.CurrencyPair)
-
-	for pairProviderName, pairs := range providerPairs {
-		for _, pair := range pairs {
-			if strings.ToUpper(pair.Quote) != config.DenomUSD {
-				// Get valid providers and use them to generate a USD-based price for this asset.
-				validProviders, err := getUSDBasedProviders(pair.Quote, providerPairs)
-				if err != nil {
-					return nil, err
-				}
-
-				// Find candles which we can use for conversion, and calculate the tvwap
-				// to find the conversion rate.
-				validCandleList := provider.AggregatedProviderCandles{}
-				for providerName, candleSet := range candles {
-					if _, ok := validProviders[providerName]; ok {
-						for base, candle := range candleSet {
-							if base == pair.Quote {
-								if _, ok := validCandleList[providerName]; !ok {
-									validCandleList[providerName] = make(map[string][]provider.CandlePrice)
-								}
-
-								validCandleList[providerName][base] = candle
-							}
-						}
-					}
-				}
-
-				if len(validCandleList) == 0 {
-					return nil, fmt.Errorf("there are no valid conversion rates for %s", pair.Quote)
-				}
-				tvwap, err := ComputeTVWAP(validCandleList)
-				if err != nil {
-					return nil, err
-				}
-
-				conversionRates[pair.Quote] = tvwap[pair.Quote]
-				requiredConversions[pairProviderName] = pair
-			}
-		}
-	}
-
-	// Convert assets to USD.
-	for provider, assetMap := range candles {
-		for asset, assetCandles := range assetMap {
-			if requiredConversions[provider].Base == asset {
-				for i := range assetCandles {
-					assetCandles[i].Price = assetCandles[i].Price.Mul(
-						conversionRates[requiredConversions[provider].Quote],
-					)
-				}
-			}
-		}
-	}
-
-	return candles, nil
-}
-
-// convertTickersToUSD converts any tickers which are not quoted in USD to USD,
-// using the conversion rates of other tickers.
-func convertTickersToUSD(
-	tickers provider.AggregatedProviderPrices,
-	providerPairs map[string][]types.CurrencyPair,
-) (provider.AggregatedProviderPrices, error) {
-	if len(tickers) == 0 {
-		return tickers, nil
-	}
-
-	conversionRates := make(map[string]sdk.Dec)
-	requiredConversions := make(map[string]types.CurrencyPair)
-
-	for pairProviderName, pairs := range providerPairs {
-		for _, pair := range pairs {
-			if strings.ToUpper(pair.Quote) != config.DenomUSD {
-				// Get valid providers and use them to generate a USD-based price for this asset.
-				validProviders, err := getUSDBasedProviders(pair.Quote, providerPairs)
-				if err != nil {
-					return nil, err
-				}
-
-				// Find valid candles, and then let's re-compute the tvwap.
-				validTickerList := provider.AggregatedProviderPrices{}
-				for providerName, candleSet := range tickers {
-					// Find tickers which we can use for conversion, and calculate the vwap
-					// to find the conversion rate.
-					if _, ok := validProviders[providerName]; ok {
-						for base, ticker := range candleSet {
-							if base == pair.Quote {
-								if _, ok := validTickerList[providerName]; !ok {
-									validTickerList[providerName] = make(map[string]provider.TickerPrice)
-								}
-
-								validTickerList[providerName][base] = ticker
-							}
-						}
-					}
-				}
-
-				if len(validTickerList) == 0 {
-					return nil, fmt.Errorf("there are no valid conversion rates for %s", pair.Quote)
-				}
-				vwap, err := ComputeVWAP(validTickerList)
-				if err != nil {
-					return nil, err
-				}
-
-				conversionRates[pair.Quote] = vwap[pair.Quote]
-				requiredConversions[pairProviderName] = pair
-			}
-		}
-	}
-
-	// Convert assets to USD.
-	for providerName, assetMap := range tickers {
-		for asset := range assetMap {
-			if requiredConversions[providerName].Base == asset {
-				assetMap[asset] = provider.TickerPrice{
-					Price: assetMap[asset].Price.Mul(
-						conversionRates[requiredConversions[providerName].Quote],
-					),
-					Volume: assetMap[asset].Volume,
-				}
-			}
-		}
-	}
-
-	return tickers, nil
-}
-
-// getUSDBasedProviders retrieves which providers for an asset have a USD-based pair,
-// given the asset and the map of providers to currency pairs.
-func getUSDBasedProviders(asset string, providerPairs map[string][]types.CurrencyPair) (map[string]struct{}, error) {
-	conversionProviders := make(map[string]struct{})
-
-	for provider, pairs := range providerPairs {
-		for _, pair := range pairs {
-			if strings.ToUpper(pair.Quote) == config.DenomUSD && strings.ToUpper(pair.Base) == asset {
-				conversionProviders[provider] = struct{}{}
-			}
-		}
-	}
-	if len(conversionProviders) == 0 {
-		return nil, fmt.Errorf("no providers have a usd conversion for this asset")
-	}
-
-	return conversionProviders, nil
 }
 
 func (o *Oracle) checkAcceptList(params oracletypes.Params) {
